Execute `scss-lint` in same dir as file
Since the content of the editor is currently written to a temporary file, the executable has no idea about where the original file is located.

`scss-lint` is a Ruby gem and often installed in a specific version for a given project, and might vary from project to project, with different, incompatible configurations.

It's trivial to write a wrapper for `scss-lint` that executes the correct version, perhaps using `bundler` and a specific version of Ruby if you use a version manager. But for this to work, we must be able to determine the original location of the project.